Module Std.Taint

Abstract taint.

This module defines several abstractions that are used to track objects as well as provide the interface to the taint tracker, that could be used for introducing and sanitizing taints or even for specifying the taint propagation policies.

module Kind : sig ... end

Each object that the engine tracks has an associated kind, that denotes a class of objects that share the same semantic properties.

module Rel : sig ... end

Relation between a value and an object that we track.

module Object : sig ... end

Each taint represents an abstract object that we would like to track. Every time a new taint is introduced we create a fresh new value, that represents the abstract object that we are tracking.

val attached : (Rel.t * Object.t * Bap_primus.Std.Primus.value) Bap_primus.Std.Primus.observation

(attached (r,o,v) occurs when the relation r is established between o and v

module Tracker : sig ... end

Taint tracker control module.

module Propagation : sig ... end

Policy based propagation.

module Gc : sig ... end