Module Taint.Kind

Each object that the engine tracks has an associated kind, that denotes a class of objects that share the same semantic properties.

In layman terms, kind allows analyses to distinguish their taints from other taints, since it introduces namespacing between different analyses so that several analyses could be ran in parallel without intersecting with each other.

type t

module Make (Machine : Bap_primus.Std.Primus.Machine.S) : sig ... end

include Core_kernel.Comparable.S_plain with type t := t and type comparator_witness = Bap_primus.Std.Primus.Value.comparator_witness

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness = Bap_primus.Std.Primus.Value.comparator_witness

val comparator : (t, comparator_witness) Base__.Comparator.comparator

val validate_lbound : min:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_ubound : max:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_bound : min:t Base__.Maybe_bound.t -> max:t Base__.Maybe_bound.t -> t Base__.Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : sig ... end

module Set : sig ... end