Module Taint.Propagation

Policy based propagation.

Although taints could be attached and detached in an arbitrary way, the framework provides a mechanism for a policy based taint propagation.

By associating a taint propagation policy with an object kind, we can specify taint propagation rules based on the kind of the object.

The policy mechanism establishes a surjection of the set of object kinds onto the set of policies, that establishes a partitioning of all objects into a disjoit set of classes, with each class defining its own rules of taint propagation.

module Policy : sig ... end