Module Gc.Conservative

Conservative Garbage Collector.

The conservative garbage collector may keep taints alive even when they become unreachable. This is an over-approximation and it is possible to devise a more precise GC, especially if soundness is not required (or not strictly required).

Taint is live if either of the following is true: 1. it is attached to a value of a variable in Env.all; 2. it is attached to any address.

The second clause gives a possibility for over-approximation, as we do not track, whether an address is reachable from the current program location. So once a tainted value is stored and the taint is attached to an address, the only way to kill this taint, is to overwrite it with another value using another store operation.

Currently, the garbage collection runs every basic block, but this may change in future.

Parameters

Machine : Bap_primus.Std.Primus.Machine.S

Signature

val init : unit -> unit Machine.t

init () component initialization function.